requirements.txt
groq
requests
python-dateutil
